translated admin pages, app.tx  half of stuff
